requirements.txt
tqdm
colorama
datetime
pyserial